+++ tests/configs/torch_spyre_tests/inductor/test_autotiling_config.yaml
@@ -1,4 +1,32 @@
 test_suite_config:
+  labels: [core, full, device_critical]
   files:
-    - path: ${TORCH_DEVICE_ROOT}/tests/inductor/test_autotiling.py
+    - path: ${TORCH_DEVICE_ROOT}/tests/inductor/test_autotiling_baseline.py
+      unlisted_test_mode: skip
+      tests:
+        - names:
+            - TestOps::test_S0_baseline_correctness.*
+            - TestOps::test_S1_pass_through.*
+            - TestOps::test_S4_boundary.*
+          mode: mandatory_success
+          tags:
+            - span_overflow__s0_s1_s4
+    - path: ${TORCH_DEVICE_ROOT}/tests/inductor/test_autotiling_spanoverflow.py
+      unlisted_test_mode: skip
+      tests:
+        - names:
+            - TestOps::test_S2_trigger_a_vocab_width_op.*
+            - TestOps::test_S6_swiglu_plan_check.*
+          mode: mandatory_success
+          tags:
+            - span_overflow__s2_s6
+    - path: ${TORCH_DEVICE_ROOT}/tests/inductor/test_autotiling_totaloverflow.py
+      unlisted_test_mode: skip
+      tests:
+        - names:
+            - TestOps::test_S9_production_stress.*
+          mode: mandatory_success
+          tags:
+            - span_overflow__s9
+    - path: ${TORCH_DEVICE_ROOT}/tests/inductor/test_autotiling_automanual.py
       unlisted_test_mode: mandatory_success
